// src/utils/types/ProfileColorPallete.ts
export type ProfileColorPalette = {
  bg: string;
  txt: string;
};

export interface LetterProfileInfo<T> {
  letter: Letter;
  palette: T;
}

export type Letter =
  | "a"
  | "b"
  | "c"
  | "d"
  | "e"
  | "f"
  | "g"
  | "h"
  | "i"
  | "j"
  | "k"
  | "l"
  | "m"
  | "n"
  | "o"
  | "p"
  | "q"
  | "r"
  | "s"
  | "t"
  | "u"
  | "v"
  | "w"
  | "x"
  | "y"
  | "z"
  | "";